menu modify, complete delete dialog

+++ src/app/features/pages/settings/settings.component.ts
@@ -1,6 +1,11 @@
 import { Component, signal } from '@angular/core';
 import { SharedModule } from '../../../shared/components/shared.module';
-import { FormBuilder, FormGroup, Validators, ɵInternalFormsSharedModule } from '@angular/forms';
+import {
+  FormBuilder,
+  FormGroup,
+  Validators,
+  ɵInternalFormsSharedModule,
+} from '@angular/forms';
 import { SettingsService } from '../../../core/services/settings.service';
 import { Employee } from 'src/app/core/models/Employee';
 
@@ -29,6 +34,8 @@ export class SettingsComponent {
     });
   }
   visible: boolean = false;
+  deleteModal: boolean = false;
+  selectedEmployee: Employee | null = null;
 
   jobTitles = [
     { label: 'Business Analyst', value: 'businessAnalyst' },
@@ -55,7 +62,7 @@ export class SettingsComponent {
     { label: 'Active', value: true },
     { label: 'Inactive', value: false },
   ];
-   
+
   showAddDialog() {
     this.visible = true;
   }
@@ -71,19 +78,26 @@ export class SettingsComponent {
     });
   }
 
-  deleteEmployee(employee: Employee) {
-    const employeeId = employee.id;
-    console.log('deleting employee with id:', employee);
-    
-    
+  openDeleteDialog(employee: Employee) {
+    this.deleteModal = true;
+    this.selectedEmployee = employee;
+  }
+
+  deleteEmployee() {
+    if (!this.selectedEmployee) {
+      console.error('No employee selected for deletion.');
+      return;
+    }
+    const employeeId = this.selectedEmployee.id;
+    console.log('deleting employee with id:', employeeId);
+
     this.settingsService.deleteEmployee(employeeId).subscribe(() => {
       this.getEmployees();
-     });
+    });
+    this.deleteModal = false;
   }
 
   editEmployee(employee: Employee) {}
-  
-    
 
   onSubmit() {
     this.settingsService.addEmployee(this.employeeForm.value).subscribe(() => {
